pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Dmitrii L <hexletdmitrii@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
django-bootstrap5 = "^24.3"
django-filter = "^24.3"
rollbar = "^1.1.0"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
